+++ src/main/java/de/maxhenkel/peek/events/DecoratedPotRenderEvents.java
@@ -9,7 +9,6 @@ import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.core.component.DataComponentPatch;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.Tag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.resources.ResourceLocation;
@@ -43,7 +42,7 @@ public class DecoratedPotRenderEvents {
         poseStack.pushPose();
         poseStack.scale(0.5F, 0.5F, 0.5F);
         poseStack.translate(0D, 4D / 16D, 0D);
-        mc.getItemRenderer().renderStatic(renderItemStack, ItemDisplayContext.GUI, light, overlay, poseStack, multiBufferSource, mc.level, 0);
+        mc.getItemRenderer().renderStatic(renderItemStack, ItemDisplayContext.FIXED, light, overlay, poseStack, multiBufferSource, mc.level, 0);
         poseStack.popPose();
         poseStack.translate(0D, -5D / 16D, 0D);
         MutableComponent text = Component.translatable("label.peek.decorated_pot_count", containedItem.getCount());
@@ -58,7 +57,7 @@ public class DecoratedPotRenderEvents {
 
     public static boolean isPotRenderStack(ItemStack stack) {
         CompoundTag tag = stack.getComponents().getOrDefault(DataComponents.CUSTOM_DATA, CustomData.EMPTY).getUnsafe();
-        return tag.contains(DECORATED_POT_ITEM_TAG, Tag.TAG_BYTE);
+        return tag.contains(DECORATED_POT_ITEM_TAG);
     }
 
     public static ItemStack createPotRenderStack(ItemStack source) {