+++ src/main/java/de/maxhenkel/peek/events/HudEvents.java
@@ -53,7 +53,7 @@ public class HudEvents {
         }
 
         Block block = blockState.getBlock();
-        ItemStack item = block.defaultBlockState().getCloneItemStack(level, blockHitResult.getBlockPos(), true); //TODO Check what the boolean does
+        ItemStack item = block.defaultBlockState().getCloneItemStack(level, blockHitResult.getBlockPos(), true);
 
         Component name = item.getHoverName();
         Font font = MC.font;