+++ src/main/java/de/maxhenkel/peek/events/ShulkerRenderEvents.java
@@ -11,7 +11,6 @@ import net.minecraft.core.Direction;
 import net.minecraft.core.component.DataComponentPatch;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.Tag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.ItemDisplayContext;
@@ -66,7 +65,7 @@ public class ShulkerRenderEvents {
                 poseStack.scale(12F / 16F, 12F / 16F, 12F / 16F);
             }
             ItemStack renderItemStack = createShulkerRenderStack(displayItem);
-            mc.getItemRenderer().renderStatic(renderItemStack, ItemDisplayContext.GUI, light, overlay, poseStack, multiBufferSource, mc.level, 0);
+            mc.getItemRenderer().renderStatic(renderItemStack, ItemDisplayContext.FIXED, light, overlay, poseStack, multiBufferSource, mc.level, 0);
             poseStack.popPose();
         }
 
@@ -88,7 +87,7 @@ public class ShulkerRenderEvents {
 
     public static boolean isShulkerRenderStack(ItemStack stack) {
         CompoundTag tag = stack.getComponents().getOrDefault(DataComponents.CUSTOM_DATA, CustomData.EMPTY).getUnsafe();
-        return tag.contains(SHULKER_ITEM_TAG, Tag.TAG_BYTE);
+        return tag.contains(SHULKER_ITEM_TAG);
     }
 
     public static ItemStack createShulkerRenderStack(ItemStack source) {